Update: Update code to match the Database ER Diagram.v3

+++ src/main/java/vn/pvhg/leaderboard/repo/PlatformRepo.java
@@ -6,7 +6,7 @@ import vn.pvhg.leaderboard.model.Platform;
 import java.util.List;
 import java.util.Optional;
 
-public interface PlatformRepo extends JpaRepository<Platform, Integer> {
+public interface PlatformRepo extends JpaRepository<Platform, Long> {
     Optional<Platform> findByPlatformType(String platformType);
 
     List<Platform> findByPlatformTypeIn(List<String> platformTypes);
