+++ src/main/java/vn/pvhg/leaderboard/repo/RoleRepo.java
@@ -7,6 +7,6 @@ import vn.pvhg.leaderboard.model.Role;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepo extends JpaRepository<Role, Integer> {
+public interface RoleRepo extends JpaRepository<Role, Long> {
     Optional<Role> findByRoleName(Role.RoleType roleName);
 }
